requirements.txt
fdb==2.0.2
python-dotenv==1.0.1
firebirdsql
passlib>=1.7.4
